+++ packages/server/src/utils/__tests__/MockDataFactory.test.ts
@@ -133,7 +133,7 @@ describe('MockDataFactory', () => {
         turn: 0,
         isActive: false,
         participants: [],
-        lairActions: null,
+        lairActions: [],
         combatLogs: [],
         createdAt: new Date('2025-01-01T00:00:00.000Z'),
         updatedAt: new Date('2025-01-01T00:00:00.000Z'),
